make examples work with doctest and move generate_test_data
